// src/interfaces/responsavel.ts
export interface IResponsavel {
  ID_USUARIO: number
  MATRICULA: string
  EMAIL: string
  RG: string
  CPF: string
  SEXO: string
  DATA_NASCIMENTO: string
  NOME_COMPLETO: string
  NOME_MAE: string
  NOME_PAI: string
  ID_CARGO: number
  CARGO: string
  ID_ORGANIZACAO: number
  ORGANIZACAO_DISPOSICAO: string
  DISPONIVEL: number
  ID_ACESSO: number
  CD_TIPO_ENVOLVIMENTO: number
}